+++ src/slogo/model/backEndInternal/commands/LessThan.java
@@ -12,6 +12,6 @@ public class LessThan<T extends Comparable<T>> implements Command<Double> {
 
   @Override
   public Double execute() {
-    return value1.compareTo(value2) < 0 ? 1.0 : 0.0;
+    return (double) (value1.compareTo(value2) < 0 ? 1 : 0);
   }
 }
